DRY up warning flags (#296)
* DRY up warning flags

* Fix core and test warning for MSVC (#299)

* Fix local MSVC warnings

* Fix core and test warning for MSVC

* address comments

* Undo prefix/suffix files

* Fix '..' include warning

* Try fixing Linux and Max warnings

* Fix Mac and Linux in a bit other way


* fix merge artifacts

* 404

* fix merge artifacts.

* "ignore" warning

* doc stub

* doxygen

* error handling

+++ sdk/core/core/test/test_json_pointer.c
@@ -39,7 +39,7 @@ void test_json_pointer() {
           break;
         }
         TEST_ASSERT(result == AZ_OK);
-        buffer[i] = code_point;
+        buffer[i] = (uint8_t)code_point;
         ++i;
       }
       az_span const b = { .begin = buffer, .size = i };
@@ -65,7 +65,7 @@ void test_json_pointer() {
           break;
         }
         TEST_ASSERT(result == AZ_OK);
-        buffer[i] = code_point;
+        buffer[i] = (uint8_t)code_point;
         ++i;
       }
       az_span const b = { .begin = buffer, .size = i };
@@ -95,7 +95,7 @@ void test_json_pointer() {
           break;
         }
         TEST_ASSERT(result == AZ_OK);
-        buffer[i] = code_point;
+        buffer[i] = (uint8_t)code_point;
         ++i;
       }
       az_span const b = { .begin = buffer, .size = i };
@@ -116,7 +116,7 @@ void test_json_pointer() {
           break;
         }
         TEST_ASSERT(result == AZ_OK);
-        buffer[i] = code_point;
+        buffer[i] = (uint8_t)code_point;
         ++i;
       }
       az_span const b = { .begin = buffer, .size = i };
